requirements.txt
fastmcp==2.10.6
pyserial==3.5 